+++ src/app/home/page.tsx
@@ -1,9 +1,17 @@
-import { Screen } from '@/screens/home';
+'use client';
+
+import TimeStamp from '@/features/TimeStamp';
 
 type Props = {};
 
 const Home = () => {
-  return <Screen />;
+  return (
+    <>
+      <div className='my-20'>
+        <TimeStamp />
+      </div>
+    </>
+  );
 };
 
 export default Home;
