+++ src/app/shift/page.tsx
@@ -1,9 +1,27 @@
-import { Screen } from '@/screens/shift';
+'use client';
+
+import { useAuthContext } from '@/context/AuthContext';
+import { ShiftCalendar } from '@/features/ShiftCalendar';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
 
 type Props = {};
 
 const Shift = () => {
-  return <Screen />;
+  const { user } = useAuthContext();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!user) router.push('/');
+  }, [user]);
+
+  return (
+    <>
+      <div className='my-20'>
+        <ShiftCalendar />
+      </div>
+    </>
+  );
 };
 
 export default Shift;
