+++ src/screens/home/index.tsx
@@ -1,13 +1,5 @@
-'use client';
-
 import TimeStamp from '@/features/TimeStamp';
 
 export const Screen: React.FC = () => {
-  return (
-    <>
-      <div className='my-20'>
-        <TimeStamp />
-      </div>
-    </>
-  );
+  return <></>;
 };
